Fix leak error in Mocha.
When running a unit test under the Mocha framework, assigning to globals
triggers the leak detector and (unless you specify the --ignore-leaks
flag) causes the test to fail. This diff fixes the declaration of R such
that it no longer triggers the leak detector. Note that this affects not
just tests directly testing this module, but any Mocha tests that test
modules with a dependency on Haversine.